Remove redundant experiment
